LayoutTests/imported/w3c/web-platform-tests/mathml/presentation-markup/mpadded/mpadded-percentage-002-expected.txt


FAIL width='200%' is interpreted as the default value assert_approx_equals: inline size expected 10 +/- 1 but got 20
FAIL height='200%' is interpreted as the default value assert_approx_equals: block size expected 50 +/- 1 but got 70
FAIL depth='200%' is interpreted as the default value assert_approx_equals: block size expected 50 +/- 1 but got 80
PASS lspace='200%' is interpreted as the default value
PASS voffset='200%' is interpreted as the default value
